fix: address issues with saving current workflow
